requirements/local.txt
asgiref==3.6.0
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.3.1
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.3
tzdata==2023.3
Unipath==1.1

pkg-resources==0.0.0
pytz==2019.3